core/services/pipeline: refactor for fuzzing; add FuzzParseETHABIArgsString; bump gonum to v0.11.0"
